requirements.txt
django==4.2
pillow
djangorestframework
markdown
django-filter
drf-yasg
djangorestframework-simplejwt
django-mail-templated
black
flake8
python-decouple
coreapi
pytest
pytest-django
faker
django-mail-templated
 django-templated-mail